Move SQL object classes into a separate directory
