+++ src/AddRecordDialog.h
@@ -1,7 +1,7 @@
 #ifndef ADDRECORDDIALOG_H
 #define ADDRECORDDIALOG_H
 
-#include "sqlitetypes.h"
+#include "sql/sqlitetypes.h"
 
 #include <QDialog>
