+++ src/sql/sqlitetypes.cpp
