+++ src/tests/CMakeLists.txt
@@ -10,7 +10,7 @@ set(TESTSQLOBJECTS_SRC
     ../sqlitedb.cpp
     ../sqlitetablemodel.cpp
     ../RowLoader.cpp
-    ../sqlitetypes.cpp
+    ../sql/sqlitetypes.cpp
     ../csvparser.cpp
     ../grammar/Sqlite3Lexer.cpp
     ../grammar/Sqlite3Parser.cpp
@@ -26,7 +26,7 @@ set(TESTSQLOBJECTS_HDR
     ../grammar/sqlite3TokenTypes.hpp
     ../grammar/Sqlite3Lexer.hpp
     ../grammar/Sqlite3Parser.hpp
-    ../sqlitetypes.h
+    ../sql/sqlitetypes.h
     ../Data.h
 )
 
@@ -94,7 +94,7 @@ set(TESTREGEX_SRC
     ../sqlitedb.cpp
     ../sqlitetablemodel.cpp
     ../RowLoader.cpp
-    ../sqlitetypes.cpp
+    ../sql/sqlitetypes.cpp
     ../grammar/Sqlite3Lexer.cpp
     ../grammar/Sqlite3Parser.cpp
     ../Settings.cpp
@@ -109,7 +109,7 @@ set(TESTREGEX_HDR
     ../grammar/sqlite3TokenTypes.hpp
     ../grammar/Sqlite3Lexer.hpp
     ../grammar/Sqlite3Parser.hpp
-    ../sqlitetypes.h
+    ../sql/sqlitetypes.h
     ../Data.h
 )
